refactor(rspack_core): move asset related struct to their own file
