+++ crates/rspack_core/src/compiler/compilation.rs
@@ -4,7 +4,6 @@ use std::{
   path::PathBuf,
   sync::{
     atomic::{AtomicBool, Ordering},
-    // time::Instant,
     Arc,
   },
 };
@@ -27,6 +26,7 @@ use tokio::sync::mpsc::error::TryRecvError;
 use tracing::instrument;
 use xxhash_rust::xxh3::Xxh3;
 
+use super::asset::{AssetInfo, CompilationAsset, CompilationAssets};
 #[cfg(debug_assertions)]
 use crate::tree_shaking::visitor::TreeShakingResult;
 use crate::{
@@ -1293,104 +1293,4 @@ impl Compilation {
       .runtime_modules
       .insert(runtime_module_identifier, module);
   }
-}
-
-pub type CompilationAssets = HashMap<String, CompilationAsset>;
-
-#[derive(Debug, Clone)]
-pub struct CompilationAsset {
-  pub source: Option<BoxSource>,
-  pub info: AssetInfo,
-}
-
-impl CompilationAsset {
-  pub fn new(source: Option<BoxSource>, info: AssetInfo) -> Self {
-    Self { source, info }
-  }
-
-  pub fn with_source(source: BoxSource) -> Self {
-    Self {
-      source: Some(source),
-      info: Default::default(),
-    }
-  }
-
-  pub fn get_source(&self) -> Option<&BoxSource> {
-    self.source.as_ref()
-  }
-
-  pub fn get_source_mut(&mut self) -> Option<&mut BoxSource> {
-    self.source.as_mut()
-  }
-
-  pub fn set_source(&mut self, source: Option<BoxSource>) {
-    self.source = source;
-  }
-
-  pub fn get_info(&self) -> &AssetInfo {
-    &self.info
-  }
-
-  pub fn get_info_mut(&mut self) -> &mut AssetInfo {
-    &mut self.info
-  }
-
-  pub fn set_info(&mut self, info: AssetInfo) {
-    self.info = info;
-  }
-}
-
-#[derive(Debug, Default, Clone)]
-pub struct AssetInfo {
-  /// if the asset can be long term cached forever (contains a hash)
-  // pub immutable: bool,
-  /// whether the asset is minimized
-  pub minimized: bool,
-  /// the value(s) of the full hash used for this asset
-  // pub full_hash:
-  /// the value(s) of the chunk hash used for this asset
-  // pub chunk_hash:
-  /// the value(s) of the module hash used for this asset
-  // pub module_hash:
-  /// the value(s) of the content hash used for this asset
-  // pub content_hash:
-  /// when asset was created from a source file (potentially transformed), the original filename relative to compilation context
-  // pub source_filename:
-  /// size in bytes, only set after asset has been emitted
-  // pub size: f64,
-  /// when asset is only used for development and doesn't count towards user-facing assets
-  pub development: bool,
-  /// when asset ships data for updating an existing application (HMR)
-  pub hot_module_replacement: bool,
-  /// when asset is javascript and an ESM
-  // pub javascript_module:
-  /// related object to other assets, keyed by type of relation (only points from parent to child)
-  pub related: AssetInfoRelated,
-}
-
-impl AssetInfo {
-  pub fn with_minimized(mut self, v: bool) -> Self {
-    self.minimized = v;
-    self
-  }
-
-  pub fn with_development(mut self, v: bool) -> Self {
-    self.development = v;
-    self
-  }
-
-  pub fn with_hot_module_replacement(mut self, v: bool) -> Self {
-    self.hot_module_replacement = v;
-    self
-  }
-
-  pub fn with_related(mut self, v: AssetInfoRelated) -> Self {
-    self.related = v;
-    self
-  }
-}
-
-#[derive(Debug, Default, Clone)]
-pub struct AssetInfoRelated {
-  pub source_map: Option<String>,
 }
